added constant deduplication

+++ benchmark/benchmark_argument_parser.c
@@ -71,7 +71,7 @@ static void benchmark_argument_parser_parse_argument(char const ** argv, option_
         if (!count) {
             benchmark_argument_parser_show_usage();
         }
-        for (size_t i = 0; i < index - 1; i++) {
+        for (int i = 0; i < index - 1; i++) {
             config_array->configs[i].executionCount = count;
         }
         *countSpecified = true;
